# Conflicts:
#	app/src/main/java/cn/atsoft/dasheng/Excel/BomController.java
#	app/src/main/java/cn/atsoft/dasheng/app/controller/StockDetailsController.java
#	app/src/main/java/cn/atsoft/dasheng/app/pojo/AllBom.java
#	app/src/main/java/cn/atsoft/dasheng/app/service/impl/ContractServiceImpl.java
#	main/src/main/resources/application-local.yml

+++ app/src/main/java/cn/atsoft/dasheng/app/pojo/AllBom.java
@@ -166,7 +166,7 @@ public class AllBom {
 
             List<Map<Long, Object>> list = new ArrayList<>();
             for (ErpPartsDetail erpPartsDetail : details) {
-                list.add(this.getBom(erpPartsDetail.getSkuId(), erpPartsDetail.getNumber() * number, erpPartsDetail.getNumber()));
+                list.add(this.getBom(erpPartsDetail.getSkuId(), (int)(erpPartsDetail.getNumber() * number), erpPartsDetail.getNumber()));
             }
 
             // 循环相加
@@ -196,7 +196,7 @@ public class AllBom {
                 SkuNumber o = (SkuNumber) map.get(aLong);
                 SkuNumber skuNumber = new SkuNumber();
                 ToolUtil.copyProperties(o, skuNumber);
-                skuNumber.setNum(o.getNum() * selfNum);
+                skuNumber.setNum((int) (o.getNum() * selfNum));
                 tmp.put(aLong, skuNumber);
             }
 
@@ -208,7 +208,7 @@ public class AllBom {
             }
             skuList.put(skuId, number + num);
             SkuNumber skuNumber = new SkuNumber();
-            skuNumber.setNum(selfNum);
+            skuNumber.setNum((int) selfNum);
             skuNumber.setSkuId(skuId);
             tmp.put(skuId, skuNumber);
         }
